Move feature functions to otherCurrency.js
